fix: avoid duplicating slugs on all charts, dashboards and spaces
