+++ packages/backend/src/models/DashboardModel/DashboardModel.ts
@@ -72,6 +72,7 @@ import { SavedSqlTableName } from '../../database/entities/savedSql';
 import { SpaceTableName } from '../../database/entities/spaces';
 import { UserTable, UserTableName } from '../../database/entities/users';
 import { DbValidationTable } from '../../database/entities/validation';
+import { generateUniqueSlug } from '../../utils/SlugUtils';
 import { SpaceModel } from '../SpaceModel';
 import Transaction = Knex.Transaction;
 
@@ -996,7 +997,11 @@ export class DashboardModel {
                     name: dashboard.name,
                     description: dashboard.description,
                     space_id: space.space_id,
-                    slug: dashboard.slug,
+                    slug: await generateUniqueSlug(
+                        trx,
+                        DashboardsTableName,
+                        dashboard.slug,
+                    ),
                 })
                 .returning(['dashboard_id', 'dashboard_uuid']);
